Extract chain explorer URL mapping.
Move explorer base URL mapping into a shared chains helper so manage splits can reuse it without inline chain constants.
